[ConfigFloat] * fix ValueError: Unknown format code 'd' for object of type 'float'
Close #3388

(cherry picked from commit ed192082750fe0b2a04fb7c72921aed903295a82)

+++ lib/python/Components/config.py
@@ -1341,8 +1341,8 @@ class ConfigFloat(ConfigSequence):
 		return int(self.value[0] * float(self.limits[1][1] + 1) + self.value[1])
 
 	def setFloatInt(self, val):
-		self.value[0] = val / float(self.limits[1][1] + 1)
-		self.value[1] = val % float(self.limits[1][1] + 1)
+		self.value[0] = int(val / float(self.limits[1][1] + 1))
+		self.value[1] = int(val % float(self.limits[1][1] + 1))
 
 	floatint = property(getFloatInt, setFloatInt)
